Starting Files
Hello!  This is the next phase of the project.  I just want to see if I
can get the FiFo fixed and the timing constraints met.  I am planning on
moving the input signal from the ADC to the MASTER_CLK before the
processing.  This may eliminate some of the timing issues and enable us
to use the second fifo on a single clock instead of two which should fix
that problem.